Fix compilation issues

+++ solana/bridge_token_factory/programs/bridge_token_factory/src/constants.rs
@@ -17,10 +17,10 @@ pub const USED_NONCES_SEED: &[u8] = b"used_nonces";
 
 
 #[constant]
-pub const USED_NONCES_PER_ACCOUNT: usize = 1024;
+pub const USED_NONCES_PER_ACCOUNT: u32 = 1024;
 
 #[constant]
-pub const USED_NONCES_ACCOUNT_SIZE: usize = 8 + (USED_NONCES_PER_ACCOUNT + 7) / 8;
+pub const USED_NONCES_ACCOUNT_SIZE: u32 = 8 + (USED_NONCES_PER_ACCOUNT + 7) / 8;
 
 #[constant]
 pub const DEFAULT_ADMIN: Pubkey = pubkey!("2ajXVaqXXpHWtPnW3tKZukuXHGGjVcENjuZaWrz6NhD4"); // TODO update this to the pubkey you can sign with
